Refactor: 상수 메시지 파일 리팩토링
