+++ src/components/template/Navbar.tsx
@@ -64,7 +64,7 @@ const Navbar = () => {
         ];
         setProducts(data);
       } catch (error) {
-        alert(MESSAGES.ERROR_PRODUCT.GET_DETAIL);
+        alert(MESSAGES.PRODUCT.ERROR_GET_DETAIL);
       } finally {
         dispatch(hideLoading());
       }
