+++ src/constants/messages.ts
@@ -1,24 +1,31 @@
 export const MESSAGES = {
-  CHECK_EMAIL: '이메일이 존재하지 않습니다.',
-  CHECK_PASSWORD: '비밀번호가 일치하지 않습니다.',
-  CHECK_EMAIL_DUPLICATE: '이미 존재하는 이메일입니다.',
-  ERROR_SIGNUP: '회원 가입 중 에러가 발생했습니다.',
-  ERROR_SIGNIN: '로그인 중 에러가 발생했습니다.',
-  ERROR_MYPAGE: '회원정보 조회 중 에러가 발생했습니다.',
-  COMPLETE_LOGIN: '로그인이 완료되었습니다.',
-  COMPLETE_SIGNUP: '회원 가입이 완료되었습니다.',
-  ERROR_CART: {
-    GET: '장바구니 조회 중 에러가 발생했습니다.',
-    CREATE: '장바구니 추가 중 에러가 발생했습니다.',
-    UPDATE: '장바구니 수정 중 에러가 발생했습니다.',
-    DELETE: '장바구니 삭제 중 에러가 발생했습니다.',
+  LOGIN: {
+    CHECK_EMAIL: '이메일이 존재하지 않습니다.',
+    CHECK_PASSWORD: '비밀번호가 일치하지 않습니다.',
+    ERROR_SIGNIN: '로그인 중 에러가 발생했습니다.',
+    COMPLETE_LOGIN: '로그인이 완료되었습니다.',
   },
-  ERROR_PRODUCT: {
-    GET_DETAIL: '상품 상세 조회 중 에러가 발생했습니다.',
+  SIGNUP: {
+    CHECK_EMAIL_DUPLICATE: '이미 존재하는 이메일입니다.',
+    ERROR_SIGNUP: '회원 가입 중 에러가 발생했습니다.',
+    COMPLETE_SIGNUP: '회원 가입이 완료되었습니다.',
   },
-  CHECK_POLICY: '약관 내용에 동의해주세요.',
-  CHECK_BUY: '신청을 완료하시겠습니까?',
-  COMPLETE_BUY: '신청이 완료되었습니다.',
-  CHECK_DELETE_CART: '선택하신 상품을 삭제하시겠습니까?',
-  COMPLETE_DELETE_CART: '삭제가 완료되었습니다.',
+  PRODUCT: {
+    ERROR_GET_DETAIL: '상품상세 조회 중 에러가 발생했습니다.',
+  },
+  CART: {
+    CHECK_DELETE: '선택하신 상품을 삭제하시겠습니까?',
+    COMPLETE_DELETE: '삭제가 완료되었습니다.',
+    COMPLETE_ADD: '장바구니에 추가되었습니다.',
+    ERROR_GET: '장바구니 조회 중 에러가 발생했습니다.',
+    ERROR_CREATE: '장바구니 추가 중 에러가 발생했습니다.',
+    ERROR_UPDATE: '장바구니 수정 중 에러가 발생했습니다.',
+    ERROR_DELETE: '장바구니 삭제 중 에러가 발생했습니다.',
+  },
+  BUY: {
+    CHECK_POLICY: '약관 내용에 동의해주세요.',
+    CHECK_BUY: '신청을 완료하시겠습니까?',
+    COMPLETE_BUY: '신청이 완료되었습니다.',
+  },
+  MYPAGE: {},
 };
