+++ src/pages/Cart.tsx
@@ -11,7 +11,7 @@ import PageTitle from '@/components/common/PageTitle';
 import { ICart } from '@/interfaces/cart';
 import Input from '@/components/common/Input';
 import { useNavigate } from 'react-router-dom';
-import ModalBox from '@/components/common/ModalBox';
+import ModalBox from '@/components/template/ModalBox';
 import { IModal } from '@/interfaces/modal';
 import { ROUTES } from '@/constants/routes';
 
@@ -30,7 +30,7 @@ const Cart = () => {
     async function getData() {
       try {
         dispatch(showLoading());
-        // const data = await getCartList();
+        // const data = await getCartList(1);
         // setCart(data);
         const data: ICart[] = [
           { id: '1', title: '직장인 신용대출', bank: '우리은행' },
@@ -44,7 +44,7 @@ const Cart = () => {
         ];
         setCart(data);
       } catch (error) {
-        alert(MESSAGES.ERROR_CART.GET);
+        alert(MESSAGES.CART.ERROR_GET);
       } finally {
         dispatch(hideLoading());
       }
@@ -61,7 +61,7 @@ const Cart = () => {
       isOpen: true,
       onClickOk: handleDeleteCart,
       onClickCancel: () => setModalState((prev) => ({ ...prev, isOpen: false })),
-      text: MESSAGES.CHECK_DELETE_CART,
+      text: MESSAGES.CART.CHECK_DELETE,
     });
   };
 
@@ -69,7 +69,7 @@ const Cart = () => {
     setModalState({
       isOpen: true,
       onClickOk: () => setModalState((prev) => ({ ...prev, isOpen: false })),
-      text: MESSAGES.COMPLETE_DELETE_CART,
+      text: MESSAGES.CART.COMPLETE_DELETE,
     });
   };
