+++ src/pages/Home.tsx
@@ -94,7 +94,7 @@ const Home = () => {
         ];
         setProducts(data);
       } catch (error) {
-        alert(MESSAGES.ERROR_PRODUCT.GET_DETAIL);
+        alert(MESSAGES.PRODUCT.ERROR_GET_DETAIL);
       } finally {
         dispatch(hideLoading());
       }
