+++ src/pages/Search.tsx
@@ -92,7 +92,7 @@ const Search = () => {
         ];
         setProducts(data);
       } catch (error) {
-        alert(MESSAGES.ERROR_PRODUCT.GET_DETAIL);
+        alert(MESSAGES.PRODUCT.ERROR_GET_DETAIL);
       } finally {
         dispatch(hideLoading());
       }
